+++ application/holder/src/service/actions/index.ts
@@ -7,4 +7,6 @@ export { handler as updates } from './updates';
 export { handler as help } from './help';
 export { handler as jumpto } from './jumpto';
 export { handler as find } from './find';
+export { handler as exportSessionState } from './export.session.state';
+export { handler as importSessionState } from './import.session.state';
 export { handler as pluginsManager } from './plugins_manager';